video增加分页修改，增加省市县

+++ app-video/src/main/java/com/jdyx/app/appvideo/controller/VideoDisplayController.java
@@ -1,10 +1,8 @@
 package com.jdyx.app.appvideo.controller;
 
 import com.alibaba.fastjson.JSON;
-import com.jdyx.app.bean.LikeInfo;
-import com.jdyx.app.bean.Page;
-import com.jdyx.app.bean.VideoDisplay;
-import com.jdyx.app.bean.VideoDisplayVo;
+import com.jdyx.app.bean.*;
+import com.jdyx.app.service.LikeInfoService;
 import com.jdyx.app.service.VideoDisplayService;
 import com.jdyx.app.util.Const;
 import com.jdyx.app.util.DistanceUtil;
@@ -24,10 +22,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.thymeleaf.util.StringUtils;
-import redis.clients.jedis.Jedis;
-import redis.clients.jedis.JedisPool;
 
-import javax.servlet.http.HttpServletResponse;
 import java.io.*;
 import java.math.BigDecimal;
 import java.text.DateFormat;
@@ -40,17 +35,18 @@ import java.util.*;
 public class VideoDisplayController {
 
     @Autowired
-    VideoDisplayService videoDisplayService;
+    LikeInfoService likeInfoService;
 
     @Autowired
-    JedisPool jedisPool;
+    VideoDisplayService videoDisplayService;
+
     /**
      *
      * @param longitude
      * @param latitude
      * @return
      */
-    @RequestMapping(value = "/getAllVideoDisplay",method = RequestMethod.POST)
+    @RequestMapping(value = "/getAllVideoDisplay",method = RequestMethod.GET)
     @ResponseBody
     @ApiOperation(value="查看所有视频")
     @ApiImplicitParams({
@@ -58,9 +54,12 @@ public class VideoDisplayController {
             @ApiImplicitParam(paramType="query", name = "pageSize", value = "每页显示记录的条数", required = false, dataType = "String"),
             @ApiImplicitParam(paramType="query", name = "latitude", value = "经度", required = false, dataType = "String"),
             @ApiImplicitParam(paramType="query", name = "longitude", value = "纬度", required = false, dataType = "String"),
-            @ApiImplicitParam(paramType="query", name = "jobId", value = "岗位id", required = false, dataType = "Integer")
+            @ApiImplicitParam(paramType="query", name = "jobId", value = "岗位id", required = false, dataType = "String")
     })
     public Object getAllVideoDisplay(Page page, BigDecimal longitude, BigDecimal latitude, Integer jobId){
+        if(jobId != null && -1 == jobId ){
+            jobId = null;
+        }
         List<VideoDisplayVo> allVideoDisplay = videoDisplayService.getAllVideoDisplayVo(jobId,page.getPageNow()*page.getPageSize(),page.getPageSize());
         if(longitude == null || latitude == null ){
             log.info("无定位信息");
@@ -70,25 +69,30 @@ public class VideoDisplayController {
                 videoDisplayVo.setDistance(distance);
             }
         }
+        //根据jobid获取总记录数
+        int totalRow = videoDisplayService.getVideoDisplayTotalByJobId(jobId);
+        //设置分页
         page.setVideoDisplayVo(allVideoDisplay);
+        page.setTotalRow(totalRow);
+        page.setTotalPage(page.getTotalRow());
         return ResultUtil.successMap(JSON.toJSON(page));
     }
 
-    @RequestMapping(value = "/getAllVideoDisplayById",method = RequestMethod.POST)
+    @RequestMapping(value = "/getAllVideoDisplayById",method = RequestMethod.GET)
     @ResponseBody
     @ApiOperation(value="查看本人视频")
     @ApiImplicitParams({
             @ApiImplicitParam(paramType="query", name = "userId", value = "用户id", required = true, dataType = "String")
     })
-    public Object getAllVideoDisplayById(Integer userId){
-        if (userId == null){
-            return ResultUtil.exceptionMap("2019","无效的参数数据");
+    public Object getAllVideoDisplayByUserId(Integer userId){
+        if (userId == null || userId == 0){
+            return ResultUtil.exceptionMap(2019,"无效的参数数据");
         }
         try {
             List<VideoDisplay> allVideoDisplay = videoDisplayService.getAllVideoDisplayById(userId);
             Object json = JSON.toJSON(allVideoDisplay);
             log.info("获取该用户所有视频 ：{}",json);
-            return ResultUtil.successMap(JSON.toJSON(allVideoDisplay));
+            return ResultUtil.successMap(json);
         }catch (Exception e){
             log.error("",e);
             return ResultUtil.errorMap();
@@ -98,16 +102,6 @@ public class VideoDisplayController {
     /**
      * 保存视频数据
      * @param videoDisplay
-     * {
-    "jobId": 22,
-    "musicAddress": "音乐地址",
-    "releaseType": "0",
-    "userId": 4,
-    "videoAddress": "http://v.jdyxqq.com/Fg4_BC95PkLiPLDgfnCBmHhU5P_o",
-    "videoDescription": "视频描述",
-    "videoPicture": "视频封面图片",
-    "videoTitle": "视频标题"
-    }
      * @return
      */
     @RequestMapping(value = "/saveVideoDisplay",method = RequestMethod.POST)
@@ -123,24 +117,36 @@ public class VideoDisplayController {
             @ApiImplicitParam(paramType="query",name = "videoPicture", value = "视频封面图片", required = true, dataType = "String"),
             @ApiImplicitParam(paramType="query",name = "videoTitle", value = "视频标题", required = true, dataType = "String")
     })
-    public Object saveVideoDisplay(VideoDisplay videoDisplay, HttpServletResponse response){
-        if(videoDisplay ==null){
-            return ResultUtil.exceptionMap("2019","无效的参数数据");
+    public Object saveVideoDisplay(VideoDisplay videoDisplay){
+        if(videoDisplay.getUserId() == null || videoDisplay.getUserId()==0){
+            return ResultUtil.exceptionMap(2019,"用户id无效");
+        }
+        if(videoDisplay.getJobId() == null || videoDisplay.getJobId()==0){
+            return ResultUtil.exceptionMap(2019,"岗位id无效");
+        }
+        if(videoDisplay.getReleaseType() == null || videoDisplay.getReleaseType()<0 || videoDisplay.getReleaseType()>1){
+            return ResultUtil.exceptionMap(2019,"短视频类型无效");
+        }
+        if(StringUtils.isEmpty(videoDisplay.getVideoAddress()) || !videoDisplay.getVideoAddress().contains(Const.QINIU_ADRESS)){
+            return ResultUtil.exceptionMap(2019,"视频地址无效");
+        }
+        if(StringUtils.isEmpty(videoDisplay.getVideoDescription())){
+            return ResultUtil.exceptionMap(2019,"视频描述无效");
+        }
+        if(StringUtils.isEmpty(videoDisplay.getVideoPicture())){
+            return ResultUtil.exceptionMap(2019,"视频封面图片无效");
+        }
+        if(StringUtils.isEmpty(videoDisplay.getVideoAddress())){
+            return ResultUtil.exceptionMap(2019,"视频标题无效");
         }
         videoDisplay.setVideoDate(new Date());
         //根据地址获取视频文件名
-        if(!videoDisplay.getVideoAddress().contains(Const.QINIU_ADRESS)){
-            return ResultUtil.exceptionMap("2019","视频地址错误");
-        }
         String bucketKey = videoDisplay.getVideoAddress().replace(Const.QINIU_ADRESS, "");
-        if(StringUtils.isEmpty(bucketKey)){
-            return ResultUtil.exceptionMap("2019","视频地址错误");
-        }
         try {
             //对视频增加水印,生成有水印的视频地址
             String videoWatermarkAddress = produceVoideoWatermark(videoDisplay.getUserId(), videoDisplay.getReleaseType(), bucketKey);
             if(StringUtils.isEmpty(videoWatermarkAddress)){
-                return ResultUtil.exceptionMap("2021","视频增加水印失败");
+                return ResultUtil.exceptionMap(2021,"视频增加水印失败");
             }
             videoDisplay.setVideoWatermarkAddress(videoWatermarkAddress);
             log.info("视频水印地址：{}",videoWatermarkAddress);
@@ -150,14 +156,33 @@ public class VideoDisplayController {
         }
         try {
             videoDisplayService.saveVideoDisplay(videoDisplay);
-            Jedis jedis = jedisPool.getResource();
-            jedis.zadd(Const.LIKE_VIDEO+videoDisplay.getId(),0,"");
-            jedis.zadd(Const.WATCH_VIDEO+videoDisplay.getId(),0,"");
             return ResultUtil.successMap("");
         }catch (Exception e){
-            return ResultUtil.exceptionMap("2020","存储数据失败");
+            return ResultUtil.exceptionMap(2020,"存储数据失败");
         }
-
+    }
+    /**
+     * 根据视频id查看个人视频
+     * @param videoId
+     * @return
+     */
+    @RequestMapping(value = "/getVideoDisplayByVideoId",method = RequestMethod.GET)
+    @ResponseBody
+    @ApiOperation(value="保存视频")
+    @ApiImplicitParams({
+            @ApiImplicitParam(paramType="query",name = "videoId", value = "视频id", required = true, dataType = "String")
+    })
+    public Object getVideoDisplayByVideoId(Integer videoId){
+        if(videoId ==null || videoId==0 ){
+            return ResultUtil.exceptionMap(2019,"视频id无效");
+        }
+        try {
+            VideoDisplay videoDisplay = videoDisplayService.getVideoDisplayByVideoId(videoId);
+            return ResultUtil.successMap(videoDisplay);
+        }catch (Exception e){
+            log.error("",e);
+            return ResultUtil.errorMap();
+        }
     }
 
     /**
@@ -172,8 +197,8 @@ public class VideoDisplayController {
             @ApiImplicitParam(paramType="query",name = "videoId", value = "视频id", required = true, dataType = "String")
     })
     public Object deleteVideoDisplay(Integer videoId){
-        if(videoId ==null){
-            return ResultUtil.exceptionMap("2019","无效的参数数据");
+        if(videoId ==null || videoId==0 ){
+            return ResultUtil.exceptionMap(2019,"无效的参数数据");
         }
         try {
             videoDisplayService.deleteVideoDisplay(videoId);
@@ -192,7 +217,7 @@ public class VideoDisplayController {
      * @return 新视频地址
      * @throws IOException
      */
-    public String produceVoideoWatermark(Integer userId,String releaseType,String bucketKey) throws IOException {
+    public String produceVoideoWatermark(Integer userId,Integer releaseType,String bucketKey) throws IOException {
         //设置账号的AK,SK
         Auth auth = Auth.create(Const.ACCESS_KEY, Const.SECRET_KEY);
 
@@ -246,21 +271,129 @@ public class VideoDisplayController {
             @ApiImplicitParam(paramType="query",name = "videoId", value = "视频id", required = true, dataType = "String")
     })
     public Object watchVideo(Integer videoId){
-        if (videoId==null){
-            return ResultUtil.exceptionMap("2019","视频id无效");
+        if (videoId==null || videoId==0){
+            return ResultUtil.exceptionMap(2019,"视频id无效");
         }
-        Jedis jedis = jedisPool.getResource();
         try {
-            //redis播放次数加1
-            jedis.incr(Const.WATCH_VIDEO+videoId);
+            //数据库修改播放次数
+            videoDisplayService.watchVideo(videoId);
             return ResultUtil.successMap("");
         } catch (Exception e) {
             e.printStackTrace();
-            return ResultUtil.exceptionMap("2020","存储数据失败");
-        } finally {
-            jedis.close();
+            return ResultUtil.exceptionMap(2020,"存储数据失败");
+        }
+    }
+    /**
+     * 视频点赞
+     * @param likeInfo
+     * @return
+     */
+    @RequestMapping(value = "/getLikeInfo",method = RequestMethod.POST)
+    @ResponseBody
+    @ApiOperation(value="视频点赞")
+    @ApiImplicitParams({
+            @ApiImplicitParam(paramType="query",name = "videoId", value = "视频id", required = true, dataType = "String"),
+            @ApiImplicitParam(paramType="query",name = "followedId", value = "点赞人id", required = true, dataType = "String"),
+    })
+    public Object getLikeInfo(LikeInfo likeInfo){
+        if (likeInfo.getVideoId()==null || likeInfo.getVideoId()==0){
+            return ResultUtil.exceptionMap(2019,"视频id无效");
+        }
+        if (likeInfo.getFollowedId()==null || likeInfo.getFollowedId()==0 ){
+            return ResultUtil.exceptionMap(2019,"点赞人id无效");
+        }
+        LikeInfo selectOne = likeInfoService.getLikeInfo(likeInfo);
+        return ResultUtil.successMap("");
+    }
+
+    /**
+     * 视频点赞
+     * @param likeInfo
+     * @return
+     */
+    @RequestMapping(value = "/likeVideo",method = RequestMethod.POST)
+    @ResponseBody
+    @ApiOperation(value="视频点赞")
+    @ApiImplicitParams({
+            @ApiImplicitParam(paramType="query",name = "videoId", value = "视频id", required = true, dataType = "String"),
+            @ApiImplicitParam(paramType="query",name = "userId", value = "短视频所属人id", required = true, dataType = "String"),
+            @ApiImplicitParam(paramType="query",name = "followedId", value = "点赞人id", required = true, dataType = "String"),
+            @ApiImplicitParam(paramType="query",name = "follName", value = "点赞人姓名", required = true, dataType = "String")
+    })
+    public Object likeVideo(LikeInfo likeInfo){
+        if (likeInfo.getVideoId()==null || likeInfo.getVideoId()==0){
+            return ResultUtil.exceptionMap(2019,"视频id无效");
+        }
+        if (likeInfo.getUserId()==null || likeInfo.getUserId()==0){
+            return ResultUtil.exceptionMap(2019,"短视频所属人id无效");
+        }
+        if (likeInfo.getFollowedId()==null || likeInfo.getFollowedId()==0 ){
+            return ResultUtil.exceptionMap(2019,"点赞人id无效");
+        }
+        if (StringUtils.isEmpty(likeInfo.getFollName())){
+            return ResultUtil.exceptionMap(2019,"点赞人姓名id无效");
+        }
+        try {
+            //存储点赞信息到数据库
+            LikeInfo selectOne = likeInfoService.getLikeInfo(likeInfo);
+            if (selectOne!=null){
+                return ResultUtil.exceptionMap(2022,"已点赞");
+            }else {
+                VideoDisplay videoDisplay = videoDisplayService.likeVideo(likeInfo.getVideoId());
+                //保存关注数据
+                likeInfo.setFollowedDate(new Date());
+                likeInfoService.saveLikeInfo(likeInfo);
+                LikeInfoVo likeInfoVo = new LikeInfoVo();
+                likeInfoVo.setVideoId(videoDisplay.getVideoId());
+                likeInfoVo.setIsLike(1);
+                likeInfoVo.setVideoLikes(videoDisplay.getVideoLikes());
+                return ResultUtil.successMap(JSON.toJSON(likeInfoVo));
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResultUtil.exceptionMap(2020,"存储数据失败");
         }
     }
 
+
+    /**
+     * 视频点赞
+     * @param likeInfo
+     * @return
+     */
+    @RequestMapping(value = "/cancelLikeVideo",method = RequestMethod.POST)
+    @ResponseBody
+    @ApiOperation(value="取消视频点赞")
+    @ApiImplicitParams({
+            @ApiImplicitParam(paramType="query",name = "videoId", value = "视频id", required = true, dataType = "String"),
+            @ApiImplicitParam(paramType="query",name = "followedId", value = "取消点赞人id", required = true, dataType = "String")
+    })
+    public Object cancelLikeVideo(LikeInfo likeInfo){
+        if (likeInfo == null || likeInfo.getVideoId() == null || likeInfo.getVideoId() == 0){
+            return ResultUtil.exceptionMap(2019,"视频id无效");
+        }
+        if (likeInfo.getFollowedId() == null){
+            return ResultUtil.exceptionMap(2019,"取消点赞人id无效");
+        }
+        try {
+            //查看点赞信息
+            LikeInfo selectOne = likeInfoService.getLikeInfo(likeInfo);
+            if (selectOne == null){
+                return ResultUtil.exceptionMap(2022,"未点赞");
+            }else {
+                VideoDisplay videoDisplay = videoDisplayService.cancelLikeVideo(likeInfo.getVideoId());
+                //删除关注信息
+                likeInfoService.deleteLikeInfo(likeInfo);
+                LikeInfoVo likeInfoVo = new LikeInfoVo();
+                likeInfoVo.setVideoId(videoDisplay.getVideoId());
+                likeInfoVo.setIsLike(0);
+                likeInfoVo.setVideoLikes(videoDisplay.getVideoLikes());
+                return ResultUtil.successMap(JSON.toJSON(likeInfoVo));
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResultUtil.errorMap();
+        }
+    }
 
 }
